+++ test/common/index.mjs
@@ -20,7 +20,6 @@ const {
   hasCrypto,
   hasIntl,
   hasIPv6,
-  hasMultiLocalhost,
   isAIX,
   isAlive,
   isDumbTerminal,
@@ -76,7 +75,6 @@ export {
   hasCrypto,
   hasIntl,
   hasIPv6,
-  hasMultiLocalhost,
   isAIX,
   isAlive,
   isDumbTerminal,
