+++ test/common/net.js
@@ -17,7 +17,17 @@ function checkSupportReusePort() {
   });
 }
 
+function hasMultiLocalhost() {
+  const { internalBinding } = require('internal/test/binding');
+  const { TCP, constants: TCPConstants } = internalBinding('tcp_wrap');
+  const t = new TCP(TCPConstants.SOCKET);
+  const ret = t.bind('127.0.0.2', 0);
+  t.close();
+  return ret === 0;
+}
+
 module.exports = {
   checkSupportReusePort,
+  hasMultiLocalhost,
   options,
 };
